feat(data-align): 实现笔记点赞增量数据处理与布隆过滤器校验
- 新增 Redis 布隆过滤器 Lua 脚本，用于校验日增量变更数据- 创建 InsertRecordMapper 及 XML 映射文件，支持笔记和用户点赞数落库
- 定义 LikeUnlikeNoteMqDTO用于 MQ 消息传输
- 配置 RedisTemplate 支持 JSON 序列化
- 修改 TableConstants 中 buildTableNameSuffix 方法参数类型为 long
- 实现 TodayNoteLikeIncrementData2DBConsumer 消费者逻辑：
  - 使用布隆过滤器去重判断 - 数据库写入操作使用事务保证原子性
  - 写入成功后更新布隆过滤器
- 更新 IntelliJ IDEA 数据源映射与 SQL 检查配置

+++ han-note-data-align/src/main/java/com/hanserwei/hannote/data/align/constant/TableConstants.java
@@ -13,7 +13,7 @@ public class TableConstants {
      * @param hashKey 哈希Keu
      * @return 表名后缀
      */
-    public static String buildTableNameSuffix(String date, int hashKey) {
+    public static String buildTableNameSuffix(String date, long hashKey) {
         // 拼接完整的表名
         return date + TABLE_NAME_SEPARATE + hashKey;
     }
